feat[STOCK-51] - Added user segmentation for Tinkoff and MOEX - Implemented error and its handling for functionality unavailable to the broker - Added migration
