+++ src/main/java/lissa/trading/user/service/controller/InternalController.java
@@ -15,7 +15,7 @@ import lissa.trading.user.service.exception.UnauthorizedException;
 import lissa.trading.user.service.feign.TinkoffAccountClient;
 import lissa.trading.user.service.page.CustomPage;
 import lissa.trading.user.service.service.UserService;
-import lissa.trading.user.service.service.creation.TempUserCreationService;
+import lissa.trading.user.service.service.creation.temp.TempUserCreationService;
 import lissa.trading.user.service.service.publisher.StatsPublisher;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -56,6 +56,12 @@ public class InternalController {
     @ApiResponse(
             responseCode = "401",
             description = "Пользователь не авторизован",
+            content = @Content(schema = @Schema(implementation = ErrorResponse.class))
+    )
+    @ApiResponse(
+            responseCode = "206",
+            description = "Пользователь успешно зарегистрирован, но" +
+                    " брокер не поддерживает получение данных",
             content = @Content(schema = @Schema(implementation = ErrorResponse.class))
     )
     @PostMapping("/register")
