+++ src/main/java/lissa/trading/user/service/dto/post/UserPostDto.java
@@ -19,6 +19,5 @@ public class UserPostDto {
     @NotNull
     @Size(min = 1, max = 50)
     private String telegramNickname;
-    @NotNull
     private String tinkoffToken;
 }
