+++ src/main/java/lissa/trading/user/service/exception/OperationUnsupportedByBrokerException.java
@@ -1,0 +1,7 @@
+package lissa.trading.user.service.exception;
+
+public class OperationUnsupportedByBrokerException extends RuntimeException {
+    public OperationUnsupportedByBrokerException(String message) {
+        super(message);
+    }
+}
