+++ src/main/java/lissa/trading/user/service/mapper/TempUserRegMapper.java
@@ -1,8 +1,10 @@
 package lissa.trading.user.service.mapper;
 
 import lissa.trading.lissa.auth.lib.dto.UserInfoDto;
+import lissa.trading.lissa.auth.lib.security.EncryptionService;
 import lissa.trading.user.service.dto.response.TempUserRegResponseDto;
 import lissa.trading.user.service.model.TempUserReg;
+import lissa.trading.user.service.service.creation.factory.SupportedBrokersEnum;
 import org.mapstruct.AfterMapping;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
@@ -21,6 +23,7 @@ public interface TempUserRegMapper {
     @Mapping(target = "externalId", ignore = true)
     @Mapping(target = "createdAt", ignore = true)
     @Mapping(target = "updatedAt", ignore = true)
+    @Mapping(target = "broker", ignore = true)
     TempUserReg toTempUserReg(UserInfoDto userInfoDto);
 
     @Mapping(target = "externalId")
@@ -30,6 +33,24 @@ public interface TempUserRegMapper {
     default void setExternalId(@MappingTarget TempUserReg tempUserReg) {
         if (tempUserReg.getExternalId() == null) {
             tempUserReg.setExternalId(UUID.randomUUID());
+        }
+    }
+
+    @AfterMapping
+    default void decryptTinkoffToken(@MappingTarget TempUserReg tempUserReg) {
+        if (tempUserReg.getTinkoffToken() != null) {
+            tempUserReg.setTinkoffToken(EncryptionService.decrypt(tempUserReg.getTinkoffToken()));
+        }
+    }
+
+    @AfterMapping
+    default void setBroker(@MappingTarget TempUserReg tempUserReg) {
+        String tinkoffToken = tempUserReg.getTinkoffToken();
+        if (tinkoffToken == null || tinkoffToken.isEmpty()) {
+            tempUserReg.setBroker(SupportedBrokersEnum.MOEX);
+        }
+        else if (tinkoffToken.startsWith("t.") && tinkoffToken.length() == 88) {
+            tempUserReg.setBroker(SupportedBrokersEnum.TINKOFF);
         }
     }
 }
