+++ src/main/java/lissa/trading/user/service/model/TempUserReg.java
@@ -2,12 +2,15 @@ package lissa.trading.user.service.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
+import lissa.trading.user.service.service.creation.factory.SupportedBrokersEnum;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -47,9 +50,12 @@ public class TempUserReg {
     @Column(name = "telegram_nickname", unique = true)
     private String telegramNickname;
 
-    @NotNull
     @Column(name = "tinkoff_token")
     private String tinkoffToken;
+
+    @Column(name = "broker")
+    @Enumerated(EnumType.STRING)
+    private SupportedBrokersEnum broker;
 
     @CreationTimestamp
     @Column(name = "created_at", nullable = false, updatable = false)
