+++ src/main/java/lissa/trading/user/service/service/creation/factory/SimpleCreationServiceFactory.java
@@ -1,0 +1,7 @@
+package lissa.trading.user.service.service.creation.factory;
+
+import lissa.trading.user.service.service.creation.UserCreationService;
+
+public interface SimpleCreationServiceFactory {
+    UserCreationService getUserCreationServiceByType(SupportedBrokersEnum supportedBroker);
+}
