+++ src/main/java/lissa/trading/user/service/service/creation/factory/SupportedBrokersEnum.java
@@ -1,0 +1,6 @@
+package lissa.trading.user.service.service.creation.factory;
+
+public enum SupportedBrokersEnum {
+    TINKOFF,
+    MOEX
+}
