+++ src/main/java/lissa/trading/user/service/service/creation/temp/TempUserCreationService.java
@@ -1,5 +1,4 @@
-package lissa.trading.user.service.service.creation;
-
+package lissa.trading.user.service.service.creation.temp;
 
 import lissa.trading.lissa.auth.lib.dto.UserInfoDto;
