+++ src/main/resources/db/changelog/09.12.2024/01_update_users_table.yaml
@@ -1,0 +1,15 @@
+databaseChangeLog:
+  - changeSet:
+      id: 09.12.2024_01
+      author: blaskowitzwork@gmail.com
+      changes:
+        - dropNotNullConstraint:
+            tableName: users
+            columnName: tinkoff_token
+            columnDataType: character varying
+        - addColumn:
+            tableName: users
+            columns:
+              - column:
+                  name: broker
+                  type: character varying(30)
