+++ src/main/resources/db/changelog/db.changelog-master.yaml
@@ -11,3 +11,6 @@ databaseChangeLog:
   - include:
       file: "07.11.2024/db.changelog-release.yaml"
       relativeToChangelogFile: true
+  - include:
+      file: "09.12.2024/db.changelog-release.yaml"
+      relativeToChangelogFile: true
