+++ src/test/java/lissa/trading/user/service/service/TinkoffUserCreationServiceImplTest.java
@@ -32,7 +32,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-class UserCreationServiceImplTest extends BaseTest {
+class TinkoffUserCreationServiceImplTest extends BaseTest {
 
     @Test
     void createUserFromTempUserReg_success() {
